Split the tests for template and views files separately under the 'tests' folder.
